BONUS: move the unit tests to a subdirectory
